+++ diagnostics/teleconnections/teleconnections/__init__.py
@@ -9,7 +9,7 @@ from .tc_class import Teleconnection
 from .tools import TeleconnectionsConfig
 from .tools import wgt_area_mean
 
-__version__ = '0.5.0'
+__version__ = '0.5.1'
 
 __all__ = ['bootstrap_teleconnections', 'build_confidence_mask',
            'station_based_index', 'regional_mean_index',
@@ -21,6 +21,7 @@ __all__ = ['bootstrap_teleconnections', 'build_confidence_mask',
            'wgt_area_mean']
 
 # Changelog
+# 0.5.1: Keeping attributes in xarray operations, final deliverable
 # 0.5.0: Bootstrap teleconnections and confidence masks functions added
 # 0.4.0: CLI is now plotting maps with the differences wrt the reference
 #        data if available